PR review update: ensure title, participants passed on update in BoardSerializer

+++ todolist/goals/serializers/board.py
@@ -54,30 +54,38 @@ class BoardSerializer(serializers.ModelSerializer):
 
     def update(self, instance, validated_data):
         owner = validated_data.pop("user")
-        new_participants = validated_data.pop("participants")
-        new_by_id = {part["user"].id: part for part in new_participants}
 
-        old_participants = instance.participants.exclude(user=owner)
         with transaction.atomic():
-            for old_participant in old_participants:
-                if old_participant.user_id not in new_by_id:
-                    old_participant.delete()
-                else:
-                    if (
-                            old_participant.role
-                            != new_by_id[old_participant.user_id]["role"]
-                    ):
-                        old_participant.role = new_by_id[old_participant.user_id][
-                            "role"
-                        ]
-                        old_participant.save()
-                    new_by_id.pop(old_participant.user_id)
-            for new_part in new_by_id.values():
-                BoardParticipant.objects.create(
-                    board=instance, user=new_part["user"], role=new_part["role"]
-                )
+            # change participants if passed
+            if validated_data.get("participants"):
+                new_participants = validated_data.pop("participants")
+                new_by_id = {part["user"].id: part for part in new_participants}
+                old_participants = instance.participants.exclude(user=owner)
 
-            instance.title = validated_data["title"]
+                # remove existing participants if not passed, change roles
+                for old_participant in old_participants:
+                    if old_participant.user_id not in new_by_id:
+                        old_participant.delete()
+                    else:
+                        if (
+                                old_participant.role
+                                != new_by_id[old_participant.user_id]["role"]
+                        ):
+                            old_participant.role = new_by_id[old_participant.user_id][
+                                "role"
+                            ]
+                            old_participant.save()
+                        new_by_id.pop(old_participant.user_id)
+
+                for new_part in new_by_id.values():
+                    BoardParticipant.objects.create(
+                        board=instance, user=new_part["user"], role=new_part["role"]
+                    )
+
+            # update title if passed
+            if validated_data.get('title'):
+                instance.title = validated_data["title"]
+
             instance.save()
 
         return instance
